Move submitProblemFactChange to fetch requests file
